+++ lib/benchmarks/controlled_gallery.dart
@@ -7,7 +7,7 @@ import 'dart:convert' show json;
 import 'dart:html' as html;
 import 'dart:math' as math;
 
-import 'perf.dart';
+import 'gallery_recorder.dart';
 import 'recorder.dart';
 
 typedef RecorderFactory = Recorder Function();
